fixed typos, separated tests and added block_number
