+++ calc/src/calc_payout.rs
@@ -122,130 +122,3 @@ impl CalcPayout {
         own_staking_payout.to_string()
     }
 }
-
-#[cfg(test)]
-mod test_payout {
-    use super::*;
-    /*
-    These tests are designed to test the accuracy of the preceding functions.
-    All the data can be retrieved and checked with polkadot-js and subscan.io.
-    NOTE: As these tests are setup, the eras should be regularly updated.
-    */
-    #[test]
-    fn kusama_era_5529_validator() {
-        /*
-        Parameters:
-            era: 5529
-            total_reward_points: 6_341_260 retrieved via api.query.staking.erasRewardPoints
-            era_payout: 792_713_971_465_885 retrieved via api.query.staking.erasValidatorReward
-            validator: CaxeCQ3JWSrZiRNyCTnE4vT8aMrX1sJDJWCXSwrEpxWkiL5
-            validator_reward_points: 3_920 retrieved via api.query.staking.erasRewardPoints
-            validator_commission: 10 % retrieved via api.query.staking.erasValidatorPrefs
-            nominator_exposure: 5_545_118_499_777 retrieved via api.query.staking.erasStakers
-            total_exposure: 9_676_879_871_438_978 retrieved via api.query.staking.erasStakers
-        */
-
-        let total_reward_points = 6_341_260u32;
-        let era_payout = String::from("792713971465885");
-
-        let params = CalcPayout::from_params(total_reward_points, &era_payout);
-
-        let validator_reward_points = 3_920u32;
-        let validator_commission = 10u32;
-        let nominator_exposure = String::from("5545118499777");
-        let total_exposure = String::from("9676879871438978");
-        let is_validator = true;
-
-        let estimated_payout = CalcPayout::calc_payout(
-            &params,
-            validator_reward_points,
-            validator_commission,
-            &nominator_exposure,
-            &total_exposure,
-            is_validator,
-        );
-
-        //https://kusama.subscan.io/event?address=CaxeCQ3JWSrZiRNyCTnE4vT8aMrX1sJDJWCXSwrEpxWkiL5&module=staking&event=rewarded&startDate=&endDate=&startBlock=&endBlock=&timeType=date&version=9430
-        let total_actual_payout = "49256160022";
-        assert_eq!(estimated_payout, total_actual_payout)
-    }
-
-    #[test]
-    fn polkadot_era_1150_validator() {
-        /*
-        Parameters:
-            era: 1150
-            total_reward_points: 22_265_020 retrieved via api.query.staking.erasRewardPoints
-            era_payout: 3_213_084_537_093_535 retrieved via api.query.staking.erasValidatorReward
-            validator: 14xKzzU1ZYDnzFj7FgdtDAYSMJNARjDc2gNw4XAFDgr4uXgp
-            validator_reward_points: 56_220 retrieved via api.query.staking.erasRewardPoints
-            validator_commission: 3 % retrieved via api.query.staking.erasValidatorPrefs
-            nominator_exposure: 4_423_101_721_494 retrieved via api.query.staking.erasStakers
-            total_exposure: 20_509_805_345_780_557 retrieved via api.query.staking.erasStakers
-        */
-
-        let total_reward_points = 22_265_020u32;
-        let era_payout = String::from("3213084537093535");
-
-        let params = CalcPayout::from_params(total_reward_points, &era_payout);
-
-        let validator_reward_points = 56_220u32;
-        let validator_commission = 3u32;
-        let nominator_exposure = String::from("4423101721494");
-        let total_exposure = String::from("20509805345780557");
-        let is_validator = true;
-
-        let estimated_payout = CalcPayout::calc_payout(
-            &params,
-            validator_reward_points,
-            validator_commission,
-            &nominator_exposure,
-            &total_exposure,
-            is_validator,
-        );
-
-        //https://polkadot.subscan.io/extrinsic/16584606-2?event=16584606-230
-        let total_actual_payout: &str = "245091889606";
-        assert_eq!(estimated_payout, total_actual_payout)
-    }
-
-    #[test]
-    fn polkadot_era_1150_nominator() {
-        /*
-        Parameters:
-            era: 1150
-            total_reward_points: 22_265_020 retrieved via api.query.staking.erasRewardPoints
-            era_payout: 3_213_084_537_093_535 retrieved via api.query.staking.erasValidatorReward
-            validator: 12MgK2Sc8Rrh6DXS2gDrt7fWJ24eGeVb23NALbZLMw1grnkL
-            nominator: 14xA7KotR6pxt3LpgdZz8BDv3fyokWnP67bBnN6tsCWn5wsF
-            validator_reward_points: 74_280 retrieved via api.query.staking.erasRewardPoints
-            validator_commission: 3 % retrieved via api.query.staking.erasValidatorPrefs
-            nominator_exposure: 4_669_514_624_960 retrieved via api.query.staking.erasStakers
-            total_exposure: 20_509_805_345_780_557 retrieved via api.query.staking.erasStakers
-        */
-
-        let total_reward_points = 22_265_020u32;
-        let era_payout = String::from("3213084537093535");
-
-        let params = CalcPayout::from_params(total_reward_points, &era_payout);
-
-        let validator_reward_points = 74_280u32;
-        let validator_commission = 2u32;
-        let nominator_exposure = String::from("4669514624960");
-        let total_exposure = String::from("20509838437005865");
-        let is_validator = false;
-
-        let estimated_payout = CalcPayout::calc_payout(
-            &params,
-            validator_reward_points,
-            validator_commission,
-            &nominator_exposure,
-            &total_exposure,
-            is_validator,
-        );
-
-        //https://polkadot.subscan.io/extrinsic/16584521-2?event=16584521-1840
-        let total_actual_payout: &str = "2391688616";
-        assert_eq!(estimated_payout, total_actual_payout)
-    }
-}
